requirements.txt
discord.py
Pillow
time